Fix tree resize problem on item collapse with horizontal selection
If an expanded tree with horizontal selection collapses the vertical bar
might be shown if the tree's width undercuts the adapter's client area
width. This fix provides the according width adjustments.

Note that there are still some minor issues regarding width calculation
on item collapse with the package explorer for example, but the vertical
bar seems not to show up anymore.