Workaround on initial bounds reconciliation on owner drawn with border
Attempt to fix #33, [ScrollableAdapter] when using EGit the dialog that displays after a push has no visible content in the top half where the tree view was.

Again, MeasurmentItem event listener cause unexpected behavior with respect to bound computations. In conjunction with borders bounds reconciliation encounters an initial scrollable location offset that let the mechanism fail and shrink the adapter to zero size. The workaround detects this offset and adjusts accordingly.