Fix #45, New Driver Definition Dialog layouts poorly on preference tab
This works around a layout problem on property sheet pages that are based on pagebook composites. The property sheet viewer calculates the tree layout only once. If this is done before the adapter mechanism isn't in place we are doomed. Hence, this approach makes an educated guess about the initialization state and filters layout events on the tree until the adapter mechanism is considered ready.