Resolve #58, redraw problem on owner drawn structured scrollables
Actually, I did not come up with an idea to debug this issue properly yet. This is because tracking down drawing problems usually suffer from some kind of Heißenberg effects - meaning that switching between the source code IDE and the UI under test during a debugging session triggers drawing events which do not occur on general usage and hence falsify the state under surveillance.

However, it seems not implausible to assume that sometimes redraw events for owner drawn virtual table or tree items might get lost due to the unexpected nested widget tree structure that comes with the scrollable adaption mechanism. This commit provides a workaround that forces a redraw on vertical scrolling explicitly.