Another attempt to fix #57 by a modification of the original workaround
This reduces the redraw suspension clamp as it caused problems with shell maximization. Unfortunately this leads to a bit of turmoil in open editors when adjusting to the correct font settings. However, in the test runs the correct font settings were successfully applied after a few moments, which seems to be improvement...