Workaround for layout problems of text viewers with ruler layout
Sometimes when opening text editors the left ruler does not show up. This commit provides a workaround that recognizes if the adapted styled text's size is (0,0). This seems to go hand in hand with the layout problem. The workaround uses a particular bound computation in that case. However, this might not be sufficient and only cures the symptom. Further observation and investigation is needed for a proper solution.