Introduction of Windows specific theme fragment
This is the first step towards resolving issue #56, Make the Clean Sheet theme compatible with the OSX version of Eclipse. Transfer the scroll bar overlay mechanism's extension declarations into a separate fragment should prevent that this functionality shows up on OSX. To keep things simple, at least for the moment, the implementations will reside in the host bundle. Note that the fragment is also loaded on GTK to allow the extension integration tests to be executed on the build server. We'll have to think about dealing with this later on...