Introduction of cocoa specific test filters
Since the scroll bar overlay tests are not related to cocoa it should be save to filter them. Of course, it would be better to keep these tests in window specific fragments. However, as this would take quite some effort and, even more important, there isn't a windows CI build available at the time being we keep them in the host bundle to be able to run at least a few of them during the Linux build.