Introduction of defensive copy for color setters in ImageDrawer
This is a workaround due to strange occasional exceptions on drawing flat scroll bar images (overlay mechanism). I wasn't able to debug the actual cause yet but I noticed that the symptom reason is a injected background/foreground color reference disposed on the outside. Hence, I switched to use defensive copy on color attributes setters. However, getters don't make use of defensive copy which is a looks a bit unbalanced. But this is necessary to avoid dangling native resource handles as the newly created copies would go out of hand and might never be disposed. Unfortunately, it would still be possible to dispose color instances once they are retrieved by a color getter. I will observer if this has an impact of the original problem.