Workaround for bounds setting problem of ScrolledCompositeAdapter
When opening the workbench's preferences dialog, the selected content is not shown properly since the scrolled composite adapter does not accept the calculated bounds due to an unexpected bounds reconciliation step. I wasn't able to find out the actual reason for this (or reproduce it in a test case for that matter), so I decided to go with a workaround that checks if the setting succeeded and repeats the application in the case of failure. Not nice, but for the time being it's the best I could come up with.