Fix #63, Fast scroll (STRG+Scroll) is not working
For some reason that I don't really understand, STRG+MouseWheel scroll up/down on a StyledText component issues a MouseWheel event to its parent adapter. However, this might have to do with the somewhat inconsistent parent child relationship given by the adapter mechanism...

Anyway, the event is picked up by the vertical FlatScrollBar control which overrides the fast scroll with different values. To avoid this problem I introduced a filter on display that stops mouse wheel event propagation on StyledTextAdapters. This seems to be safe from my current point of view. However, if against my expectations this workaround will cause any trouble it might be feasible to flag the mouse wheel handler of the FlatScrollBar control.

But for the time being I would like to avoid implementing such a hack into the low level control to solve this higher level issues.