Workaround for layout problem of StyledTextAdapters in TextViewers
When lazy loading parts with TextViewers in the workbench the location of the according StyledTextAdapter is often not computed correctly. It seems that the layout reconciliation somehow prevents the adapter location to be set correctly in this scenario. Repeating the setting of the location a second time, however, does the trick. Since neither I'm sure why this doesn't work correctly when setting the location only once nor I'm able to reproduce the problem in a test case I leave this for the time being with the resetting mechanism as workaround in place.